I hate GitHub, update on time table
Changing timetable, only started the first section on the road to a
functioning timetable.
Also adding new features like a comment section that appears with the
rest of the information when you hover over a class on the timetable.
Also adding the ability to have different background colours to each
class on the timetable to make it pop more.