BUG FIXES
Change structure of timetable, finalised functionality now just have to
do some house keeping.

Its come to my attention that when a window opens its location on the
computer screen is weirdly placed so to view the full window you have
to drag it into place. I need to put location changes on the load up
functions of each page so is fully visible and no faffing is needed